Adding .ignore file -Extension

+++ Data/Entities/ParkingTransaction.cs
@@ -6,7 +6,7 @@ namespace Parking_System_API.Data.Entities
 {
     public class ParkingTransaction
     {   [Key, Required, ForeignKey("FK_Participant_RelatedTo")]
-        public int ParticipantId { get; set; }
+        public long ParticipantId { get; set; }
         [Key, Required, ForeignKey("FK_Vehicle_RelatedTo")]
         public string PlateNumberId { get; set; }
         [Key, Required, ForeignKey("FK_Hardware_RelatedTo")]
