+++ Data/Repositories/ParticipantR/IParticipantRepository.cs
@@ -13,6 +13,6 @@ namespace Parking_System_API.Data.Repositories.ParticipantR
         // SystemUsers
         Task<Participant[]> GetAllParticipants(bool getVehicles = false, bool getTransactions = false);
         Task<Participant> GetParticipantAsyncByEmail(string email, bool getVehicles = false, bool getTransactions = false);
-        Task<Participant> GetParticipantAsyncByID(int id, bool getVehicles = false, bool getTransactions = false);
+        Task<Participant> GetParticipantAsyncByID(long id, bool getVehicles = false, bool getTransactions = false);
     }
 }
